requirements.txt
pygame
coloredlogs
jupyter
huggingface_hub
Pillow
picovoice
pvrecorder
pvporcupine
google-cloud-speech
pyaudio
together
openai
